+++ ui/src/onboarding/components/CompletionStep.tsx
@@ -15,7 +15,7 @@ import {
   QuickstartDashboardCreationSuccess,
   QuickstartDashboardCreationError,
 } from 'src/shared/copy/notifications'
-import {QUICKSTART_DASHBOARD_NAME} from 'src/onboarding/constants'
+import {ossMetricsTemplate} from 'src/templates/constants/defaultTemplates'
 
 // APIs
 import {getDashboards} from 'src/organizations/apis'
@@ -31,7 +31,7 @@ import {
   Columns,
   Grid,
 } from '@influxdata/clockface'
-import {DashboardTemplate, Organization} from 'src/types'
+import {Organization} from 'src/types'
 import {Dashboard, ScraperTargetRequest} from '@influxdata/influx'
 import {OnboardingStepProps} from 'src/onboarding/containers/OnboardingWizard'
 import {QUICKSTART_SCRAPER_TARGET_URL} from 'src/dataLoaders/constants/pluginConfigs'
@@ -165,23 +165,10 @@ class CompletionStep extends PureComponent<Props> {
       this.props.notify(QuickstartScraperCreationError)
     }
     try {
-      const templatesEntries = await client.templates.getAll(this.props.orgID)
-      const templatesToGet = templatesEntries.filter(t => {
-        return (
-          t.meta.name.toLowerCase() == QUICKSTART_DASHBOARD_NAME.toLowerCase()
-        )
-      })
-      const pendingTemplates = templatesToGet.map(t =>
-        client.templates.get(t.id)
+      await createDashboardFromTemplateAJAX(
+        ossMetricsTemplate(),
+        this.props.orgID
       )
-      const templates = await Promise.all(pendingTemplates)
-      const pendingDashboards = templates.map(t =>
-        createDashboardFromTemplateAJAX(
-          t as DashboardTemplate,
-          this.props.orgID
-        )
-      )
-      await Promise.all(pendingDashboards)
       this.props.notify(QuickstartDashboardCreationSuccess)
     } catch (err) {
       this.props.notify(QuickstartDashboardCreationError)