+++ ui/src/onboarding/constants/index.ts
@@ -1,1 +1,1 @@
-export const QUICKSTART_DASHBOARD_NAME = 'Local Metrics-Template'
+export const QUICKSTART_DASHBOARD_NAME = 'InfluxDB 2.0 OSS Metrics'